pyproject.toml
[project]
name = "lenticularis"
version = "0.1.0"
description = "Weather Data tracking with focus on Ruleset for certain Locations and the surrounding Weather Data"
authors = [
    {name = "Raphael Eymann, r+lenti@eymanns.ch"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = "^3.11"
dependencies = [
    "fastapi (>=0.124.4,<0.125.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "influxdb-client (>=1.49.0,<2.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "apscheduler (>=3.11.1,<4.0.0)",
    "sqlalchemy (>=2.0.45,<3.0.0)",
    "alembic (>=1.17.2,<2.0.0)",
    "pyproj (>=3.7.0,<4.0.0)",
    "python-multipart (>=0.0.20,<0.1.0)",
    "python-jose[cryptography] (>=3.3.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "email-validator (>=2.0.0,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "black (>=25.12.0,<26.0.0)",
    "ruff (>=0.14.9,<0.15.0)"
]
